// src/design-system/tokens/build/react-native/tokens.d.ts
export const colorBgAccentHigh: string;
export const colorBgAccentHighAccented: string;
export const colorBgGenerativeStrong: string;
export const colorBgGenerativeHigh: string;
export const colorBgGenerativeHighAccented: string;
export const colorBgCarbyDefault: string;
export const colorBgCarbyDefaultAccent: string;
export const colorBgCarbyHighContrast: string;
export const colorFgNeutralPrimary: string;
export const colorFgNeutralSecondary: string;
// ...
export const colorFgAccentPrimary: string;
// ...
export const dimensionSpaceBetweenRepeatingSm: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
/** Gap between cards, etc. */
export const dimensionSpaceBetweenRepeatingMd: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceBetweenRelatedSm: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceBetweenRelatedMd: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceBetweenSeparatedSm: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const dimensionSpaceAroundMd: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceAroundLg: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const textFontSizeDisplayLg: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const textFontSizeHeadingMd: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const textFontSizeHeadingLg: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const textFontSizeBodyXs: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const textFontSizeBodySm: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const textLineHeightDisplayLg: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const textLineHeightHeadingMd: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const textLineHeightHeadingLg: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const textLineHeightBodyXs: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const textLineHeightBodySm: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
// ...
export const textFontWeightDisplayBold: string;
export const textFontWeightSubheading: string;
export const textFontWeightHeadingMedium: string;
export const textFontWeightHeadingBold: string;
export const textFontWeightBodyRegular: string;
export const textFontWeightBodyMedium: string;
export const textFontWeightBodyBold: string;
export const textFontWeightLabelMedium: string;
export const textFontWeightEyebrowMedium: string;
export const textLetterSpacingDisplay: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};